refactor: move sources/sinks to pipeline and remove logging from init files
