--- src/fabricflow/copy/__init__.py
@@ -1,4 +1,0 @@
-import logging
-from logging import Logger
-
-logger: Logger = logging.getLogger(__name__)
