+++ src/fabricflow/pipeline/sinks/__init__.py
@@ -1,0 +1,12 @@
+from .lakehouse_table import LakehouseTableSink
+from .parquet_file import ParquetFileSink
+from .base import BaseSink
+from .types import SinkType
+
+
+__all__: list[str] = [
+    "LakehouseTableSink",
+    "ParquetFileSink",
+    "BaseSink",
+    "SinkType",
+]
