+++ src/fabricflow/pipeline/sinks/types.py
@@ -1,12 +1,6 @@
 from enum import Enum
 
 
-class SourceType(Enum):
-    """Enum for different types of data sources."""
-
-    SQL_SERVER = "SQLServer"
-
-
 class SinkType(Enum):
     """Enum for different types of data sinks."""
