+++ src/fabricflow/pipeline/sources/__init__.py
@@ -1,0 +1,9 @@
+from .types import SourceType
+from .base import BaseSource
+from .sql_server import SQLServerSource
+
+__all__: list[str] = [
+    "SourceType",
+    "BaseSource",
+    "SQLServerSource",
+]
