+++ src/fabricflow/pipeline/sources/types.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class SourceType(Enum):
+    """Enum for different types of data sources."""
+
+    SQL_SERVER = "SQLServer"
